% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-freshen.R
\name{better_exists}
\alias{better_exists}
\title{Better exists()}
\usage{
better_exists(
  x,
  where = -1,
  mode = "any",
  inherits = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{character vector length 1 or more, indicating the
object names, with or without package prefix.}

\item{where, mode, inherits}{arguments passed to \code{base::exists()}.
Note that arguments \code{envir} and \code{frame} use the defaults,
and therefore may not be compatible with using input \code{x}
with more than one value.}

\item{...}{additional arguments are passed to \code{base::exists()}.

#' @examples
exists("exists", where="package:base")

better_exists("base::exists")}
}
\value{
logical vector with length equal to \code{length(x)}. Note that
when a package prefix is supplied, when the package is not on
the search path this function returns \code{FALSE}, and does not
throw an error.
}
\description{
Better exists()
}
\concept{jam utility functions

This function is a lightweight enhancement of \code{base::exists()}
that accepts a package prefix in the object name, for example
\code{"base::exists"}.

This function recognizes a package prefix \code{"packagename::"}
and uses it to determine the correct value for argument \code{where}.
If the package is not present in \code{search()} using the
form \code{"package:packagename"} then an error is thrown.
Otherwise if the package is on the search path, this
function simply calls \code{base::exists(x, where=posnum, ...)}.

If the input \code{x} does not contain a package prefix, then
this function simply calls \code{base::exists(x)}
for the default behavior.}
