% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-freshen.R
\name{better_get}
\alias{better_get}
\title{Better get()}
\usage{
better_get(x, pos = -1L, mode = "any", inherits = TRUE, ...)
}
\arguments{
\item{pos, mode, inherits}{arguments passed to \code{base::get()}. Note
that argument \code{envir} is not passed to \code{base::get()}, since
it is typically defined dynamically by that function.}

\item{...}{additional arguments are passed to \code{base::get()}. The
only additional argument is \code{envir} which is not recommended,
but is here for compatibility with the base functionality.}

\item{character}{string with an R object name, with or without
an R package prefix. Note that only one value is recognized.}
}
\value{
the R object found. If not object is found an error results.
}
\description{
Better get()
}
\examples{
get("get", pos="package:base")

better_get("base::get")

}
\concept{jam utility functions

This function is a lightweight enhancement of \code{base::get()}
that accepts a package prefix in the object name, for example
\code{"base::exists"}.

This function recognizes a package prefix \code{"packagename::"}
and uses it to determine the correct value for argument \code{where}.
If the package is not present in \code{search()} using the
form \code{"package:packagename"} then an error is thrown.
Otherwise if the package is on the search path, this
function simply calls \code{base::get(x, pos=posnum, ...)}.

If the input \code{x} does not contain a package prefix, then
this function simply calls \code{base::get()}
for the default behavior.}
