% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-convenience.R
\name{freshenGenes2}
\alias{freshenGenes2}
\title{Freshen gene annotations using Bioconductor annotation data}
\usage{
freshenGenes2(
  x,
  ann_lib = c("", "org.Hs.eg.db"),
  try_list = c("SYMBOL2EG", "ACCNUM2EG", "ALIAS2EG"),
  final = c("SYMBOL", "GENENAME"),
  split = "[ ]*[,/;]+[ ]*",
  sep = ",",
  handle_multiple = c("first_try", "first_hit", "all", "best_each"),
  empty_rule = c("empty", "original", "na"),
  include_source = FALSE,
  protect_inline_sep = TRUE,
  intermediate = "intermediate",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{character} vector or \code{data.frame} with one or most columns
containing gene symbols.}

\item{ann_lib}{\code{character} vector indicating the name or names of the
Bioconductor annotation library to use when looking up
gene nomenclature.}

\item{try_list}{\code{character} vector indicating one or more names of
annotations to use for the input gene symbols in \code{x}. The
annotation should typically return the Entrez gene ID, usually
given by \code{'2EG'} at the end of the name. For example \code{SYMBOL2EG}
will be used with ann_lib \code{"org.Hs.eg.db"} to produce annotation
name \code{"org.Hs.egSYMBOL2EG"}. Note that when the \code{'2EG'} form of
annotation does not exist (or another suitable suffix defined in
argument \code{"revmap_suffix"} in \code{get_anno_db()}), it will be derived
using \code{AnnotationDbi::revmap()}. For example if \code{"org.Hs.egALIAS"}
is requested, but only \code{"org.Hs.egALIAS2EG"} is available, then
\code{AnnotationDbi::revmap(org.Hs.egALIAS2EG)} is used to create the
equivalent of \code{"org.Hs.egALIAS"}.}

\item{final}{\code{character} vector to use for the final conversion
step. When \code{final} is \code{NULL} no conversion is performed.
When \code{final} contains multiple values, each value is returned
in the output. For example, \code{final=c("SYMBOL","GENENAME")} will
return a column \code{"SYMBOL"} and a column \code{"GENENAME"}.}

\item{split}{\code{character} value used to separate delimited values in \code{x}
by the function \code{base::strsplit()}. The default will split values
separated by comma \verb{,} semicolon \verb{;} or forward slash \code{/}, and will
trim whitespace before and after these delimiters.}

\item{sep}{\code{character} value used to concatenate multiple entries in
the same field. The default \code{sep=","} will comma-delimit multiple
entries in the same field.}

\item{handle_multiple}{\code{character} value indicating how to handle multiple
values: \code{"first_hit"} will query each column of \code{x} until it finds the
first possible returning match, and will ignore all subsequent possible
matches for that row in \code{x}. For example, if one row in \code{x} contains
multiple values, only the first match will be used. \code{"first_try"}
will return the first match from \code{try_list} for all columns in \code{x}
that contain a match. For example, if one row in \code{x} contains two
values, the first match from \code{try_list} using one or both columns in
\code{x} will be maintained. Subsequent entries in \code{try_list} will not be
attempted for rows that already have a match. \code{"all"} will return all
possible matches for all entries in \code{x} using all items in \code{try_list}.}

\item{empty_rule}{\code{character} value indicating how to handle entries which
did not have a match, and are therefore empty: \code{"original"} will use
the original entry as the output field; \code{"empty"} will leave the
entry blank.}

\item{include_source}{\code{logical} indicating whether to include a column
that shows the colname and source matched. For example, if column
\code{"original_gene"} matched \code{"SYMBOL2EG"} in \code{"org.Hs.eg.db"} there
will be a column \code{"found_source"} with value
\code{"original_gene.org.Hs.egSYMBOL2EG"}.}

\item{protect_inline_sep}{\code{logical} indicating whether to
protect inline characters in \code{sep}, to prevent them from
being used to split single values into multiple values.
For example, \code{"GENENAME"} returns the full gene name, which
often contains comma \code{","} characters. These commas do
not separate multiple separate values, so they should not be
used to split a string like \code{"H4 clustered histone 10, pseudogene"}
into two strings \code{"H4 clustered histone 10"} and \code{"pseudogene"}.}

\item{intermediate}{\code{character} string with colname in \code{x} that
contains intermediate values. These values are expected from output
of the first step in the workflow, for example \code{"SYMBOL2EG"}
returns Entrez gene values, so if the input \code{x} already contains
some of these values in a column, assign that colname to
\code{intermediate}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}
}
\description{
Freshen gene annotations using Bioconductor annotation data
}
\details{
This function is a convenient extension of \code{freshenGenes()}
that adds \code{GENENAME} to the default value for
\code{final=c("SYMBOL", "GENENAME")}.
It therefore returns two (\code{2}) annotation columns by default,
the gene symbol, and the long gene name.
}
\seealso{
Other genejam: 
\code{\link{freshenGenes}()},
\code{\link{freshenGenes3}()},
\code{\link{get_anno_db}()},
\code{\link{is_empty}()}
}
\concept{genejam}
