% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-freshen.R
\name{imget}
\alias{imget}
\title{Case-insensitive mget()}
\usage{
imget(
  x,
  envir = as.environment(-1L),
  mode = "any",
  ifnotfound = NA,
  inherits = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{character vector of object names.}

\item{envir, mode, ifnotfound, inherits}{arguments are passed to
\code{base:mget()}.}
}
\value{
named \code{list} of objects found, or \code{NA} for objects
that are not found.
}
\description{
Case-insensitive mget()
}
\details{
This function is a lightweight wrapper around \code{base::mget()}
(and generics) that intends to allow case-insensitive matching.
It does so by converting all keys to lowercase, matching
lowercase input to these lowercase keys, then using the original
keys in native \code{base::mget()}.

One small change from \code{base::mget()} is the default
argument \code{ifnotfound=NA}.

This function secretly runs \code{mget()} using the unique lowercase
input values \code{x}, to reduce the number of queries. This implementation
is designed to help with extremely long and potentially highly duplicated
input values in \code{x}, in which case the change greatly reduces the time
to return results.

Note: This function returns the first matching lowercase
key, with the direct assumption that keys will not be duplicated
after converting to lowercase. Should this assumption become a
problem, please provide feedback and we will change the method
accordingly.

Note: For unknown reasons, the R method dispatch was not
behaving properly for objects of class \code{"AnnDbBimap"}, presumably
because the generic functions \code{AnnotationDbi::ls()} and
\code{AnnotationDbi::mget()} were written for class \code{"Bimap"}.
So when the input \code{envir} class contains \code{"Bimap"} the
direct function \code{AnnotationDbi::keys()} is called, and if it
fails for some reason, \code{AnnotationDbi::ls()} is called,
thes \code{AnnotationDbi::mget()} is called, otherwise the generic
\code{ls()} or \code{mget()} is called.
}
\concept{jam utility functions}
