% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-convenience.R
\name{is_empty}
\alias{is_empty}
\title{Test if vector elements are empty}
\usage{
is_empty(x, ...)
}
\arguments{
\item{x}{\code{vector} that may contain \code{NA} values.}

\item{...}{additional arguments are ignored.}
}
\description{
Test if vector elements are empty
}
\details{
This function simply checks if values in a vector are
\code{NA} or \code{""} with \code{nchar(x) == 0}.

For \code{factor} input, the values are coerced with \code{as.character()}.
It might be slightly faster to test factor levels then to
apply to the full vector.

Todo: Make this function work with \code{list} input, so it
requires all elements to be \code{is_empty()}.
}
\examples{
x1 <- c("A", "", NA, "B,C");
is_empty(x1)

is_empty(factor(x1))

}
\seealso{
Other genejam: 
\code{\link{freshenGenes2}()},
\code{\link{freshenGenes3}()},
\code{\link{freshenGenes}()},
\code{\link{get_anno_db}()}
}
\concept{genejam}
