% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-freshen.R
\name{lgsub}
\alias{lgsub}
\title{Pattern replacement in a list of character vectors}
\usage{
lgsub(pattern, replacement, x, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, useBytes = FALSE, ...)
}
\arguments{
\item{pattern, replacement, ignore.case, perl, fixed, useBytes}{all
arguments are passed to \code{base::gsub()} after \code{x} is converted
to a character vector.}

\item{x}{\code{list} object that contains character vectors.}

\item{...}{additional arguments are ignored.}
}
\description{
Pattern replacement in a list of character vectors
}
\details{
This function is a simple wrapper around \code{base::gsub()} except
it operates on a list.

Note that this function assumes the input data contains vectors
and not embedded list objects.
}
\examples{

x <- list(a=c("A", "B"), b=c("C,D"));
lgsub(",", "!:!", x)

}
\concept{jam list functions}
