% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-freshen.R
\name{freshenGenes}
\alias{freshenGenes}
\title{Freshen gene annotations using Bioconductor annotation data}
\usage{
freshenGenes(
  x,
  ann_lib = c("", "org.Hs.eg.db"),
  try_list = c("SYMBOL2EG", "ACCNUM2EG", "ALIAS2EG"),
  final = c("SYMBOL"),
  split = "[ ]*[,/;]+[ ]*",
  revert_split = TRUE,
  sep = ",",
  handle_multiple = c("first_try", "first_hit", "all", "best_each"),
  empty_rule = c("empty", "original", "na"),
  include_source = FALSE,
  protect_inline_sep = TRUE,
  intermediate = "ENTREZID",
  ignore.case = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{character} vector or \code{data.frame} with one or most columns
containing gene symbols.}

\item{ann_lib}{\code{character} vector indicating the name or names of the
Bioconductor annotation library to use when looking up
gene nomenclature.}

\item{try_list}{\code{character} vector indicating one or more names of
annotations to use for the input gene symbols in \code{x}. The
annotation should typically return the Entrez gene ID, usually
given by \code{'2EG'} at the end of the name. For example \code{SYMBOL2EG}
will be used with ann_lib \code{"org.Hs.eg.db"} to produce annotation
name \code{"org.Hs.egSYMBOL2EG"}. Note that when the \code{'2EG'} form of
annotation does not exist (or another suitable suffix defined in
argument \code{"revmap_suffix"} in \code{get_anno_db()}), it will be derived
using \code{AnnotationDbi::revmap()}. For example if \code{"org.Hs.egALIAS"}
is requested, but only \code{"org.Hs.egALIAS2EG"} is available, then
\code{AnnotationDbi::revmap(org.Hs.egALIAS2EG)} is used to create the
equivalent of \code{"org.Hs.egALIAS"}.}

\item{final}{\code{character} vector to use for the final conversion
step. When \code{final} is \code{NULL} no conversion is performed.
When \code{final} contains multiple values, each value is returned
in the output. For example, \code{final=c("SYMBOL","GENENAME")} will
return a column \code{"SYMBOL"} and a column \code{"GENENAME"}.}

\item{split}{\code{character} value used to separate delimited values in \code{x}
by the function \code{base::strsplit()}. The default will split values
separated by comma \verb{,} semicolon \verb{;} or forward slash \code{/}, and will
trim whitespace before and after these delimiters.}

\item{revert_split}{\code{logical} default TRUE, whether to revert the
internal split with argument \code{split}, which causes columns with
delimited values to be split into single value per column.
When \code{revert_split=TRUE} the columns are re-combined to match
the original data. Note that values are combined using \code{sep}
as delimiter, which may not be identical to the input data.}

\item{sep}{\code{character} value used to concatenate multiple entries in
the same field. The default \code{sep=","} will comma-delimit multiple
entries in the same field.}

\item{handle_multiple}{\code{character} value indicating how to handle multiple
values:
\itemize{
\item \code{"first_hit"} will query each column of \code{x} until it finds the
first possible returning match, and will ignore all subsequent possible
matches for that row in \code{x}. For example, if one row in \code{x} contains
multiple values, only the first match will be used.
\item \code{"first_try"}
will return the first match from \code{try_list} for all columns in \code{x}
that contain a match. For example, if one row in \code{x} contains two
values, the first match from \code{try_list} using one or both columns in
\code{x} will be maintained. Subsequent entries in \code{try_list} will not be
attempted for rows that already have a match.
\item \code{"all"} will return all
possible matches for all entries in \code{x} using all items in \code{try_list}.
}}

\item{empty_rule}{\code{character} value indicating how to handle entries which
did not have a match, and are therefore empty: \code{"original"} will use
the original entry as the output field; \code{"empty"} will leave the
entry blank.}

\item{include_source}{\code{logical} indicating whether to include a column
that shows the colname and source matched. For example, if column
\code{"original_gene"} matched \code{"SYMBOL2EG"} in \code{"org.Hs.eg.db"} there
will be a column \code{"found_source"} with value
\code{"original_gene.org.Hs.egSYMBOL2EG"}.}

\item{protect_inline_sep}{\code{logical} indicating whether to
protect inline characters in \code{sep}, to prevent them from
being used to split single values into multiple values.
For example, \code{"GENENAME"} returns the full gene name, which
often contains comma \code{","} characters. These commas do
not separate multiple separate values, so they should not be
used to split a string like \code{"H4 clustered histone 10, pseudogene"}
into two strings \code{"H4 clustered histone 10"} and \code{"pseudogene"}.}

\item{intermediate}{\code{character} string with colname in \code{x} that
contains intermediate values. These values are expected from output
of the first step in the workflow, for example \code{"SYMBOL2EG"}
returns Entrez gene values, so if the input \code{x} already contains
some of these values in a column, assign that colname to
\code{intermediate}.}

\item{ignore.case}{\code{logical} indicating whether to use
case-insensitive matching when \code{ignore.case=TRUE}, otherwise
the default \code{ignore.case=FALSE} will perform default \code{mget()}
which requires the upper and lowercase characters are
an identical match. When \code{ignore.case=TRUE} this function
calls \code{genejam::imget()}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}
}
\value{
\code{data.frame} with one or more columns indicating the input
data, then a column \code{"intermediate"} containing the Entrez gene ID
that was matched, then one column for each item in \code{final},
by default \code{"SYMBOL"}.
}
\description{
Freshen gene annotations using Bioconductor annotation data
}
\details{
This function takes a vector or \code{data.frame} of gene symbols,
and uses Bioconductor annotation methods to find the most current
official gene symbol.

The annotation process runs in two basic steps:
\enumerate{
\item \strong{Convert the input gene to Entrez gene ID} (\code{intermediate}).
\item \strong{Convert Entrez gene ID to official gene symbol} (\code{final}).
}

The \code{intermediate} depends upon the annotation used, but in the
default condition, most Bioconductor gene-centric annotation sources
are focused on \code{ENTREZID} as the primary key for integration.

During Step 1 above, values are successively queried to obtain
a successful conversion to \code{intermediate}, the specific logic for which
is controlled by \code{handle_multiple}. However, once the criteria
are met which produces valid \code{intermediate}, the value or values
in the \code{intermediate} are then used for Step 2 above.
\subsection{Step 1. Convert to Entrez gene ID}{

The first step uses an ordered list of annotations,
with the assumption that the first match is usually the best,
and most specific. By default, the order is:
\itemize{
\item \code{"org.Hs.egSYMBOL2EG"} -- almost always 1-to-1 match
\item \code{"org.Hs.egACCNUM2EG"} -- mostly a 1-to-1 match
\item \code{"org.Hs.egALIAS2EG"} -- sometimes a 1-to-1 match, sometimes 1-to-many
}

When multiple Entrez gene ID values are matched, they are all
retained. See argument \code{handle_multiple} for custom options.
}

\subsection{Step 2. Use Entrez gene ID to return official annotation}{

The second step converts the Entrez gene ID (or multiple IDs)
to the official gene symbol, by default using \code{"org.Hs.egSYMBOL"}.

The second step may optionally include multiple annotation types,
each of which will be returned. Some common examples:
\itemize{
\item \code{"org.Hs.egSYMBOL"} -- official Entrez gene symbol
\item \code{"org.Hs.egALIAS"} -- set of recognized aliases for an Entrez gene.
\item \code{"org.Hs.egGENENAME"} -- official Entrez long gene name
}

For each step, the annotation matched can be returned, as an audit
trail to see which annotation was available for each input entry.

Note that if the input data already contains Entrez gene ID
values, you can define that colname with argument \code{intermediate}.
}

\subsection{Pre-existing intermediate values}{

When input data contains a column that matches argument \code{intermediate},
the intention is for pre-existing values to be honored and maintained.
As a result, pre-existing non-empty values in \code{intermediate} are kept
without being updated. For example, in a \code{data.frame} that contains
colnames 'gene' and 'intermediate', and values are present in both
columns for a given row, the value in 'intermediate' is used without
also querying the value in 'gene' for that row.
}

\subsection{Case-insensitive search}{

Case-insensitive search is particularly useful in non-human
organisms because they often use mixed-case gene symbols.
This support is enabled with the argument \code{ignore.case=TRUE}.
In our benchmark tests it appears to add roughly
0.1 seconds per annotation in \code{try_list}, regardless of the
number of input entries in \code{x}. It seems reasonable that the
cost may increase with annotation size, although in practice
the time dependency does not seem to be notably affected by
annotation size.

This time cost is related to the process of converting all \code{keys()}
in the annotation to lowercase, plus relatively minimal time to
handle scenarios where multiple keys in mixed case (capitalization)
are converted to the same lowercase key, for example 'CAL', 'Cal', 'cal'
are all recognized as 'cal', and all associated connections are combined.
This secondary step was added in version 0.0.18.900. Prior to that
version, only the first such matching lowercase result was used.

For the vast majority of gene-based queries involving human, this
step is not necessary and produces identical results. The reason is
that human gene symbols are entirely UPPERCASE.
}
}
\examples{
if (suppressPackageStartupMessages(require(org.Hs.eg.db))) {
   cat("\nBasic usage\n");
   print(freshenGenes(c("APOE", "CCN2", "CTGF")));
}

if (suppressPackageStartupMessages(require(org.Hs.eg.db))) {
   ## Optionally show the annotation source matched
   cat("\nOptionally show the annotation source matched\n");
   print(freshenGenes(c("APOE", "CCN2", "CTGF"), include_source=TRUE));
}

if (suppressPackageStartupMessages(require(org.Hs.eg.db))) {
   ## Show comma-delimited genes
   cat("\nInput genes are comma-delimited\n");
   print(freshenGenes(c("APOE", "CCN2", "CTGF", "CCN2,CTGF")));
}

if (suppressPackageStartupMessages(require(org.Hs.eg.db))) {
   ## Optionally include more than SYMBOL in the output
   cat("\nCustom output to include SYMBOL, ALIAS, GENENAME\n");
   print(freshenGenes(c("APOE", "HIST1H1C"),
      final=c("SYMBOL", "ALIAS", "GENENAME")));
}

if (suppressPackageStartupMessages(require(org.Hs.eg.db))) {
   ## More advanced, match affymetrix probesets
   if (requireNamespace("hgu133plus2.db", quietly=TRUE) &&
      suppressPackageStartupMessages(require("hgu133plus2.db"))) {
      cat("\nAdvanced example including Affymetrix probesets.\n");
      affy_df <- freshenGenes(
         c("227047_x_at", "APOE", "HIST1H1D", "NM_003166,U08032"),
         include_source=TRUE,
         try_list=c("hgu133plus2ENTREZID", "REFSEQ2EG",
            "SYMBOL2EG", "ACCNUM2EG", "ALIAS2EG"),
         final=c("SYMBOL", "GENENAME"))
      print(affy_df)
   }
}

}
\seealso{
Other genejam: 
\code{\link{freshenGenes2}()},
\code{\link{freshenGenes3}()},
\code{\link{get_anno_db}()},
\code{\link{is_empty}()}
}
\concept{genejam}
