% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genejam-freshen.R
\name{get_anno_db}
\alias{get_anno_db}
\title{Get annotation database or environment}
\usage{
get_anno_db(x, revmap_suffix = "2EG", verbose = FALSE, ...)
}
\arguments{
\item{x}{character name of an annotation object, or an annotation
object itself.}

\item{revmap_suffix}{character string indicting the expected suffix
that can be used to create reverse-mapped annotation data, for
example the suffix \code{"2EG"} is used to indicate that annotation
returns Entrez gene. When annotation does not contain this suffix,
the annotation is reverse-mapped using \code{AnnotationDbi::revmap()}.}

\item{verbose}{logical indicating whetheer to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Get annotation database or environment
}
\details{
This function is a simple wrapper function that takes either an
annotation data name, for example from the \code{AnnotationDbi} package,
or an annotation object, and returns the annotation object.

In the event the annotation object must be derived using
\code{AnnotationDbi::revmap()}, then that process is performed, and the
reverse mapped annotation object is returned.
}
\seealso{
Other genejam: 
\code{\link{freshenGenes}()}
}
\concept{genejam}
